Conflicts:
	src/server/authserver/Server/AuthSession.cpp
	src/server/game/Entities/Player/Player.cpp
	src/server/game/Server/WorldSocket.cpp

+++ src/server/game/Entities/Creature/Creature.cpp
@@ -2283,7 +2283,7 @@ void Creature::AllLootRemovedFromCorpse()
     if (loot.loot_type == LOOT_SKINNING)
         m_corpseRemoveTime = now;
     else
-        m_corpseRemoveTime = now + m_corpseDelay * decayRate;
+        m_corpseRemoveTime = now + uint32(m_corpseDelay * decayRate);
 
     m_respawnTime = m_corpseRemoveTime + m_respawnDelay;
 }
